requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
fastapi==0.115.0
graphql-core==3.2.4
h11==0.14.0
idna==3.10
libcst==1.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-multipart==0.0.10
PyYAML==6.0.2
rich==13.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
strawberry-graphql==0.243.1
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
